# openapi/favorX.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: favorX API
  description: "A list of the currently provided Interfaces to interact with the boson, implementing file operations and sending messages"
security:
  - { }
...
servers:
  - url: "http://{apiRoot}:{port}/v1"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local favorX node main API
      port:
        default: "1633"
        description: Service port provided in favorX node config
  - url: "http://{apiRoot}:{port}"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local favorX node main API
      port:
        default: "1633"
        description: Service port provided in favorX node config
paths:
...
  "/file":
    post:
      summary: "Upload file or a collection of files"
      description: "In order to upload a collection, user can send a multipart request with all the files populated in the form data with appropriate headers.\n\n
              User can also upload a tar file along with the collection header. This will upload the tar file after extracting the entire directory structure.\n\n
              If the collection header is absent, all requests (including tar files) are considered as single file uploads.\n\n
              A multipart request is treated as a collection regardless of whether the collection header is present. This means in order to serve single files
              uploaded as a multipart request, the index-document header should be used with the name of the file."
      tags:
        - File
        - Collection
      parameters:
        - $ref: "favorXCommon.yaml#/components/parameters/PinParameter"
        - $ref: "favorXCommon.yaml#/components/parameters/EncryptParameter"
        - $ref: "favorXCommon.yaml#/components/parameters/ContentTypePreserved"
        - $ref: "favorXCommon.yaml#/components/parameters/CollectionParameter"
        - $ref: "favorXCommon.yaml#/components/parameters/IndexDocumentParameter"
        - $ref: "favorXCommon.yaml#/components/parameters/ErrorDocumentParameter"
        - $ref: "favorXCommon.yaml#/components/parameters/ReferenceLinkParameter"
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary
          application/x-tar:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Ok
          headers:
            "etag":
              $ref: "favorXCommon.yaml#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "favorXCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "favorXCommon.yaml#/components/responses/400"
        "403":
          $ref: "favorXCommon.yaml#/components/responses/403"
        "500":
          $ref: "favorXCommon.yaml#/components/responses/500"
        default:
          description: Default response
    get:
      summary: "Page the file list"
      description: ""
      tags:
        - File
      parameters:
        - in: query
          name: filter
          schema:
            type: string
          required: false
          example: '[ { "key": "FieldName","term": "eq","value": "false" } ]'
        - in: query
          name: page
          schema:
            type: string
          required: true
          example: '{ "pageNum": 1, "pageSize": 10 }'
        - in: query
          name: sort
          schema:
            type: string
          required: false
          example: '{ "key": "FieldName", "order": "desc" }'
      responses:
        "200":
          description: "A list of file"
          content:
            application/json:
              schema:
                $ref: "favorXCommon.yaml#/components/schemas/PageResponse"
        "400":
          $ref: "favorXCommon.yaml#/components/responses/400"
        "404":
          $ref: "favorXCommon.yaml#/components/responses/404"
        "500":
          $ref: "favorXCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/file/{reference}":
    get:
      summary: "Get file or index document from a collection of files"
      tags:
        - File
        - Collection
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "favorXCommon.yaml#/components/schemas/BosonReference"
          required: true
          description: Boson address of content
        - $ref: "favorXCommon.yaml#/components/parameters/RecoveryTargetsParameter"
      responses:
        "200":
          description: Ok
          headers:
            "recovery-targets":
              $ref: "favorXCommon.yaml#/components/headers/RecoveryTargets"
          #           "feed-index":
          #             $ref: "favorXCommon.yaml#/components/headers/FeedIndex"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "favorXCommon.yaml#/components/responses/400"
        "404":
          $ref: "favorXCommon.yaml#/components/responses/404"
        "500":
          $ref: "favorXCommon.yaml#/components/responses/500"
        default:
          description: Default response
    delete:
      summary: "Delete file or a collection of files"
      tags:
        - File
        - Collection
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "favorXCommon.yaml#/components/schemas/BosonReference"
          required: true
          description: Boson address of content
      responses:
        "200":
          description: Ok
        "400":
          $ref: "favorXCommon.yaml#/components/responses/400"
        "404":
          $ref: "favorXCommon.yaml#/components/responses/404"
        "500":
          $ref: "favorXCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/file/{reference}/{path}":
    get:
      summary: "Get referenced file from a collection of files"
      tags:
        - Collection
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "favorXCommon.yaml#/components/schemas/BosonReference"
          required: true
          description: Boson address of content
        - in: path
          name: path
          schema:
            type: string
          required: true
          description: Path to the file in the collection.
        - $ref: "favorXCommon.yaml#/components/parameters/RecoveryTargetsParameter"
      responses:
        "200":
          description: Ok
        "400":
          $ref: "favorXCommon.yaml#/components/responses/400"
        "404":
          $ref: "favorXCommon.yaml#/components/responses/404"
        "500":
          $ref: "favorXCommon.yaml#/components/responses/500"
        default:
          description: Default response
...